requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
matplotlib
opencv-python-headless